sql: Move Backend API definition to a base class
